python/requirements/development.txt
attrs>=19.2.0
autodocsumm==0.1.13 #Pinned as errors on 0.2.0
biopython
black
breathe
codecov
coverage
flake8
h5py>=2.6.0
jsonschema>=3.0.0
kastore
msgpack>=1.0.0
msprime
networkx
newick
numpy
packaging
portion
pre-commit
pyparsing 
pysam
pytest
pytest-cov
pytest-xdist
python_jsonschema_objects
PyVCF
sphinx
sphinx-argparse
sphinx-issues
sphinx_rtd_theme
svgwrite>=1.1.10
xmlunittest
